Implemented the AccountCreator class which allows a user to sign up for this service. Modified the module-info.java file to allow Gson to access the classes specified in the 'opens' statements and to allow java.swing and java.swt to work specified by the requires java.desktop statement.
Added default constructors to the java classes that will simulate a song object.

BUG: Exception in thread "main" java.lang.reflect.InaccessibleObjectException: Unable to make field private final java.time.LocalDate java.time.LocalDateTime.date accessible: module java.base does not "opens java.time" to module com.google.gson

For some reason we can't create an account without this exception happening. This is due to Gson not being able to access the LocalDataTime object we have in the Playlist class. For now we'll have the playlistCreationDateTime field commented until we can figure out a solution.